Code clean-up, add some translate

+++ database/migrations/2022_06_12_115211_create_notes_table.php
@@ -15,7 +15,7 @@ class CreateNotesTable extends Migration
     {
         Schema::create('notes', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id')->index();
+            $table->unsignedBigInteger('user_id')->index();
             $table->text('title');
             $table->text('description');
             $table->timestamp('date');
